+++ cpp/tests/stream_compaction/apply_boolean_mask_tests.cpp
@@ -167,6 +167,53 @@ TEST_F(ApplyBooleanMask, FixedPointColumnTest)
   CUDF_TEST_EXPECT_TABLES_EQUAL(expect_cudf_table_view, tableView);
 }
 
+TEST_F(ApplyBooleanMask, FixedPointLargeColumnTest)
+{
+  cudf::size_type num_rows = 10000;
+
+  using decimal32_wrapper = cudf::test::fixed_point_column_wrapper<int32_t>;
+  using decimal64_wrapper = cudf::test::fixed_point_column_wrapper<int64_t>;
+
+  std::vector<int32_t> dec32_data(num_rows);
+  std::vector<int64_t> dec64_data(num_rows);
+  std::vector<bool> mask_data(num_rows);
+
+  cudf::test::UniformRandomGenerator<int32_t> rng32(-10000000, 10000000);
+  cudf::test::UniformRandomGenerator<int64_t> rng64(-1000000000000, 1000000000000);
+  cudf::test::UniformRandomGenerator<bool> rbg;
+  std::generate(dec32_data.begin(), dec32_data.end(), [&rng32]() { return rng32.generate(); });
+  std::generate(dec64_data.begin(), dec64_data.end(), [&rng64]() { return rng64.generate(); });
+  std::generate(mask_data.begin(), mask_data.end(), [&rbg]() { return rbg.generate(); });
+
+  decimal32_wrapper col32(dec32_data.begin(), dec32_data.end(), numeric::scale_type{-3});
+  decimal64_wrapper col64(dec64_data.begin(), dec64_data.end(), numeric::scale_type{-10});
+  cudf::table_view cudf_table_in_view{{col32, col64}};
+
+  cudf::test::fixed_width_column_wrapper<bool> bool_filter(mask_data.begin(), mask_data.end());
+  cudf::column_view bool_filter_col(bool_filter);
+
+  std::unique_ptr<cudf::table> filteredTable =
+    cudf::apply_boolean_mask(cudf_table_in_view, bool_filter_col);
+  cudf::table_view tableView = filteredTable->view();
+
+  std::vector<int32_t> expect_dec32_data;
+  std::vector<int64_t> expect_dec64_data;
+  for (int i = 0; i < mask_data.size(); ++i) {
+    if (mask_data[i]) {
+      expect_dec32_data.push_back(dec32_data[i]);
+      expect_dec64_data.push_back(dec64_data[i]);
+    }
+  }
+
+  decimal32_wrapper expect_col32(
+    expect_dec32_data.begin(), expect_dec32_data.end(), numeric::scale_type{-3});
+  decimal64_wrapper expect_col64(
+    expect_dec64_data.begin(), expect_dec64_data.end(), numeric::scale_type{-10});
+  cudf::table_view expect_cudf_table_view{{expect_col32, expect_col64}};
+
+  CUDF_TEST_EXPECT_TABLES_EQUAL(expect_cudf_table_view, tableView);
+}
+
 TEST_F(ApplyBooleanMask, NoNullInput)
 {
   cudf::test::fixed_width_column_wrapper<int32_t> col(